+++ packages/data/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./types";
 export * from "./opportunity-index";
+export * from "./hybrid-scoring";
 export * from "./opportunity-index-examples";
 export * from "./metric-utils";
 export * from "./boundaries";