+++ packages/geo/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./color-scales";
 export * from "./geography-cameras";
+export * from "./discovery-flyover";
 export * from "./us-map";